+++ src/server/js/methods.js
@@ -42,38 +42,44 @@ Meteor.methods({
         return await AccountsCore.s.byUsername( instanceName, username, options );
     },
 
+    // insert a new account - this is a fallback for not-users collections which do not provide a createAccount() hook
+    async 'pwix.AccountsCore.m.createAccount'( instanceName, userDoc, requesterId ){
+        check( instanceName, Match.NonEmptyString );
+        check( userDoc, Object );
+        check( requesterId, Match.NonEmptyStringject );
+        try {
+            return await AccountsCore.s.createAccount( instanceName, userDoc, requesterId );
+        } catch( e ){
+            logger.error( e );
+            return { reason: e.error };
+        }
+    },
+
+    // insert a new account in the 'users' collection
+    //  a special method to be sure the account is created on server side, thus do not auto connect
+    async 'pwix.AccountsCore.m.createUserAsync'( createOpts ){
+        return await Accounts.createUserAsync( createOpts );
+    },
+
     // delete an account
-    async 'pwix.AccountsCore.m.deleteAccount'( instanceName, user, options={} ){
+    async 'pwix.AccountsCore.m.deleteAccount'( instanceName, user, requesterId ){
         check( instanceName, Match.NonEmptyString );
-        check( user, Match.OneOf( Match.NonEmptyString, Match.ObjectIncludeingKeys({ _id: Match.NonEmptyString })));
-        check( options, Object );
-        return await AccountsCore.s.deleteAccount( instanceName, user, options );
-    },
-
-    // insert a new account - this is a fallback for not-users collections which do not provide a createUser() hook
-    async 'pwix.AccountsCore.m.insertAccount'( instanceName, userDoc, options={} ){
-        check( instanceName, Match.NonEmptyString );
-        check( userDoc, Object );
-        check( options, Object );
-        return await AccountsCore.s.insertAccount( instanceName, userDoc, options );
+        check( user, Match.OneOf( Match.NonEmptyString, Match.ObjectIncluding({ _id: Match.NonEmptyString })));
+        check( requesterId, Match.NonEmptyString );
+        try {
+            return await AccountsCore.s.deleteAccount( instanceName, user, requesterId );
+        } catch( e ){
+            logger.error( e );
+            return { reason: e.error };
+        }
     },
 
     // update an account
-    async 'pwix.AccountsCore.m.updateAccount'( instanceName, userDoc, origDoc=null, options={} ){
+    async 'pwix.AccountsCore.m.updateAccount'( instanceName, userDoc, requesterId, opts={} ){
         check( instanceName, Match.NonEmptyString );
         check( userDoc, Object );
-        check( origDoc, Match.OneOf( null, Object ));
-        check( options, Object );
-        return await AccountsCore.s.updateAccount( instanceName, userDoc, origDoc, options );
-    },
-
-    // update an account
-    //  do not care of any permission here
-    async 'pwix.AccountsCore.m.updateByQuery'( instanceName, selector, modifier, options={} ){
-        check( instanceName, Match.NonEmptyString );
-        check( selector, Object );
-        check( modifier, Object );
-        check( options, Object );
-        return await AccountsCore.s.updateByQuery( instanceName, selector, modifier, options, this.userId );
+        check( requesterId, Match.NonEmptyString );
+        check( opts, Object );
+        return await AccountsCore.s.updateAccount( instanceName, userDoc, requesterId, opts );
     }
 });